GH-2829 refactor to use componentEffect for sentry initialization
